+++ greenfield_python_sdk/blockchain_client.py
@@ -1,5 +1,4 @@
-import base64
-from typing import Optional
+from typing import List, Optional
 
 import aiohttp
 from betterproto import Casing
@@ -20,8 +19,6 @@ from greenfield_python_sdk.blockchain.virtual_group import VirtualGroup
 from greenfield_python_sdk.config import NetworkConfiguration
 from greenfield_python_sdk.key_manager import KeyManager
 from greenfield_python_sdk.models.broadcast import BroadcastMode
-from greenfield_python_sdk.models.eip712_messages.storage.bucket_url import CREATE_BUCKET, MIGRATE_BUCKET
-from greenfield_python_sdk.models.eip712_messages.storage.object_url import CREATE_OBJECT
 from greenfield_python_sdk.models.transaction import BroadcastOption
 from greenfield_python_sdk.protos.cosmos.base.v1beta1 import Coin
 from greenfield_python_sdk.protos.cosmos.crypto.secp256k1 import PubKey
@@ -35,7 +32,7 @@ from greenfield_python_sdk.protos.cosmos.tx.v1beta1 import (
     Tx,
     TxBody,
 )
-from greenfield_python_sdk.sign_utils import get_signature
+from greenfield_python_sdk.utils.sign_utils import encode_sp_approval_message, get_signatures
 
 
 class BlockchainClient:
@@ -82,8 +79,8 @@ class BlockchainClient:
 
     async def build_tx(
         self,
-        message,
-        type_url: str,
+        messages,
+        type_url: List[str],
         fee: Optional[Fee] = None,
     ) -> Tx:
         if not self.key_manager:
@@ -96,20 +93,18 @@ class BlockchainClient:
                 gas_limit=2000,
                 payer=self.key_manager.address,
             )
-        if type_url == CREATE_BUCKET or type_url == CREATE_OBJECT:  # TODO: Move to other place
-            sp_approval = base64.b64decode(message.primary_sp_approval.sig)
-            message.primary_sp_approval.sig = bytes(sp_approval)
 
-        if type_url == MIGRATE_BUCKET:
-            sp_approval = base64.b64decode(message.dst_primary_sp_approval.sig)
-            message.dst_primary_sp_approval.sig = bytes(sp_approval)
+        messages = encode_sp_approval_message(messages, type_url)
+        wrapped_message = []
+        for i in range(len(messages)):
+            wrapped_message.append(
+                AnyMessage(
+                    type_url=type_url[i],
+                    value=bytes(messages[i]),
+                )
+            )
 
-        wrapped_message = AnyMessage(
-            type_url=type_url,
-            value=bytes(message),
-        )
-
-        body = TxBody(messages=[wrapped_message])
+        body = TxBody(messages=wrapped_message)
         auth_info = AuthInfo(
             signer_infos=[
                 SignerInfo(
@@ -161,14 +156,14 @@ class BlockchainClient:
 
     async def build_tx_from_message(
         self,
-        message,
-        type_url: str,
+        messages,
+        type_url: List[str],
         fee: Optional[Fee] = None,
         broadcast_option: Optional[BroadcastOption] = None,
     ):
-        tx = await self.build_tx(message=message, type_url=type_url, fee=fee)
+        tx = await self.build_tx(messages=messages, type_url=type_url, fee=fee)
 
-        signature_pre = await get_signature(self.key_manager, tx, message, self.chain_id, broadcast_option)
+        signature_pre = await get_signatures(self.key_manager, tx, messages, self.chain_id, broadcast_option)
         tx.signatures = [signature_pre]
         try:
             simulation = await self.simulate_tx(tx)
@@ -182,19 +177,19 @@ class BlockchainClient:
             tx.auth_info.fee.gas_limit = 20000000
             tx.auth_info.fee.amount[0].amount = "100000000000000000"
 
-        signature = await get_signature(self.key_manager, tx, message, self.chain_id, broadcast_option)
+        signature = await get_signatures(self.key_manager, tx, messages, self.chain_id, broadcast_option)
         tx.signatures = [signature]
         return tx
 
     async def broadcast_message(
         self,
-        message,
-        type_url: str,
+        messages,
+        type_url: List[str],
         fee: Optional[Fee] = None,
         broadcast_option: Optional[BroadcastOption] = None,
     ):
         tx = await self.build_tx_from_message(
-            message=message,
+            messages=messages,
             type_url=type_url,
             fee=fee,
             broadcast_option=broadcast_option,