+++ tests/e2e/greenfield/test_bucket.py
@@ -308,6 +308,40 @@ async def test_buy_quota_for_bucket():
 @pytest.mark.requires_config
 @pytest.mark.tx
 @pytest.mark.slow
+async def test_create_bucket_with_tag():
+    config = get_account_configuration()
+    key_manager = KeyManager(private_key=config.private_key)
+    async with GreenfieldClient(network_configuration=network_configuration, key_manager=key_manager) as client:
+        await client.async_init()
+
+        tags = ResourceTags(tags=[ResourceTagsTag(key="tag1", value="first_tag")])
+        bucket_name = "".join(random.choices(string.ascii_lowercase + string.digits, k=random.randint(5, 11)))
+        sp = (await client.blockchain_client.get_active_sps())[0]
+        tx_hash = await client.bucket.create_bucket(
+            bucket_name,
+            primary_sp_address=sp["operator_address"],
+            opts=CreateBucketOptions(
+                charged_read_quota=100, visibility=VisibilityType.VISIBILITY_TYPE_PRIVATE, tags=tags
+            ),
+        )
+        assert tx_hash
+        assert len(tx_hash) == 64
+        assert isinstance(tx_hash, str)
+        await client.basic.wait_for_tx(hash=tx_hash)
+
+        head_bucket = await client.bucket.get_bucket_head(bucket_name)
+        assert head_bucket
+        assert head_bucket.bucket_name == bucket_name
+        assert head_bucket.tags == tags
+
+        tx_hash = await client.bucket.delete_bucket(bucket_name)
+        assert tx_hash
+        await client.basic.wait_for_tx(hash=tx_hash)
+
+
+@pytest.mark.requires_config
+@pytest.mark.tx
+@pytest.mark.slow
 async def test_create_bucket_and_set_tag():
     config = get_account_configuration()
     key_manager = KeyManager(private_key=config.private_key)